Remove abstract protocols and move private interfaces
