+++ src/readonly_fs_tools/_internal/__init__.py
@@ -1,0 +1,1 @@
+"""Internal implementation classes - not part of public API."""
