+++ src/readonly_fs_tools/grep.py
@@ -4,8 +4,8 @@ from typing import List
 
 from pydantic import BaseModel, Field
 
-from ._defaults import FilesystemPathEnumerator, StreamingRegexSearcher
-from ._protocols import PathEnumerator, RegexSearcher
+from ._internal.path_enumerator import FilesystemPathEnumerator
+from ._internal.regex_searcher import StreamingRegexSearcher
 from .budget import BudgetExceeded, OutputBudget
 from .common import FileContent, GlobPattern, RegexPattern
 from .sandbox import Sandbox
@@ -22,7 +22,9 @@ class Grepper:
     """Safe content searcher with sandbox constraints."""
 
     def __init__(
-        self, path_enum: PathEnumerator, regex_searcher: RegexSearcher
+        self,
+        path_enum: FilesystemPathEnumerator,
+        regex_searcher: StreamingRegexSearcher,
     ) -> None:
         self.path_enum = path_enum
         self.regex_searcher = regex_searcher
